requirements.txt
# ================================================================
# AIMScribe AI Backend - Python Dependencies
# ================================================================

# ================================================================
# Core Framework (Flask for legacy, FastAPI for async)
# ================================================================
flask>=3.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydub>=0.25.1
pyngrok>=7.0.0

# ================================================================
# LangChain Framework
# ================================================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-experimental>=0.0.40
langgraph>=0.0.20

# ================================================================
# Azure OpenAI
# ================================================================
openai>=1.10.0

# ================================================================
# Vector Database (RAG)
# ================================================================
chromadb>=0.4.0
tiktoken>=0.5.0

# ================================================================
# PostgreSQL Database
# ================================================================
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# ================================================================
# Template Engine
# ================================================================
jinja2>=3.1.0

# ================================================================
# Configuration
# ================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ================================================================
# Utilities
# ================================================================
redis>=5.0.0
requests>=2.31.0

# ================================================================
# Storage & Queue
# ================================================================
minio>=7.2.0
redis>=5.0.0

# ================================================================
# Development/Testing (Optional)
# ================================================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.12.0
# flake8>=6.1.0
